reply_bot/requirements.txt
BeautifulSoup4
emoji
google-generativeai
pandas
pyperclip
pytz
selenium
snscrape
webdriver-manager 
psutil
python-dotenv
PyYAML
